# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perpustakaan",
    "description": "Perpustakaan API Collection",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3880/"
    }
  ],
  "paths": {
...
    "/datamember": {
      "post": {
        "tags": [
          "Datamember"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatamemberAdd"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Datamember"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatamemberUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Datamember"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatamemberDelete"
              }
            }
          }
        }
      }
    },
...
    "/datamember/{id}": {
      "get": {
        "tags": [
          "Datamember"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/kategoribuku": {
      "post": {
        "tags": [
          "Kategoribuku"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KategoribukuAdd"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Kategoribuku"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KategoribukuUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Kategoribuku"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KategoribukuDelete"
              }
            }
          }
        }
      }
    },
...
    "/kategoribuku/{id}": {
      "get": {
        "tags": [
          "Kategoribuku"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KategoribukuAdd": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "kategoribuku label"
          }
        }
      },
      "KategoribukuUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "kategoribuku label"
          },
          "id_kategori": {
            "type": "number",
            "example": 0
          }
        }
      },
      "KategoribukuDelete": {
        "type": "object",
        "properties": {
          "id_kategori": {
            "type": "number",
            "example": 0
          }
        }
      },
...
      "DatamemberAdd": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "datamember label"
          }
        }
      },
      "DatamemberUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "datamember label"
          },
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
      "DatamemberDelete": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
...
    }
  }
}